// frontend/src/openapi/generated/model/create-memo-input.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Zete
 * 메모 서비스
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { Tag } from './tag';

/**
 * 
 * @export
 * @interface CreateMemoInput
 */
export interface CreateMemoInput {
    /**
     * 
     * @type {string}
     * @memberof CreateMemoInput
     */
    'title'?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateMemoInput
     */
    'content'?: string;
    /**
     * 
     * @type {boolean}
     * @memberof CreateMemoInput
     */
    'isImportant'?: boolean;
    /**
     * 
     * @type {number}
     * @memberof CreateMemoInput
     */
    'cateId'?: number;
    /**
     * 
     * @type {Array<Tag>}
     * @memberof CreateMemoInput
     */
    'tags'?: Array<Tag>;
}